MAGECLOUD-4076: Refactoring service EOL validation.
