+++ src/Config/Validator/Deploy/ServiceEol.php
@@ -5,9 +5,11 @@
  */
 declare(strict_types=1);
 
-namespace Magento\MagentoCloud\Config\Validator\Deploy\ServiceEol;
+namespace Magento\MagentoCloud\Config\Validator\Deploy;
+
 
 use Magento\MagentoCloud\Config\Validator;
+use Magento\MagentoCloud\Filesystem\FileSystemException;
 use Magento\MagentoCloud\Service\EolValidator as EOLValidator;
 use Magento\MagentoCloud\Config\ValidatorInterface;
 use Magento\MagentoCloud\Service\ServiceMismatchException;
@@ -15,12 +17,12 @@ use Magento\MagentoCloud\Service\ServiceMismatchException;
 /**
  * Class to check if services approaching their EOLs.
  */
-class Notices implements ValidatorInterface
+class ServiceEol implements ValidatorInterface
 {
     /**
-     * Define validation level.
+     * @var int
      */
-    private const ERROR_LEVEL = ValidatorInterface::LEVEL_NOTICE;
+    private $errorLevel;
 
     /**
      * @var Validator\ResultFactory
@@ -35,29 +37,37 @@ class Notices implements ValidatorInterface
     /**
      * @param Validator\ResultFactory $resultFactory
      * @param EOLValidator $eolValidator
+     * @param int $errorLevel
      */
     public function __construct(
         Validator\ResultFactory $resultFactory,
-        EOLValidator $eolValidator
+        EOLValidator $eolValidator,
+        int $errorLevel
     ) {
         $this->resultFactory = $resultFactory;
         $this->eolValidator = $eolValidator;
+        $this->errorLevel = $errorLevel;
     }
 
     /**
      * Get the defined services and versions and check for their EOLs by error level.
      *
      * @return Validator\ResultInterface
+     * @throws FileSystemException
      */
     public function validate(): Validator\ResultInterface
     {
         try {
             $errors = $this->eolValidator->validateServiceEol();
 
-            if ($errors && array_key_exists(self::ERROR_LEVEL, $errors)) {
+            $message = $this->errorLevel == ValidatorInterface::LEVEL_WARNING ?
+                'Some services have passed EOL.' :
+                'Some services are approaching EOL.';
+
+            if ($errors && array_key_exists($this->errorLevel, $errors)) {
                 return $this->resultFactory->error(
-                    'Some services are approaching EOL.',
-                    implode(PHP_EOL, $errors[self::ERROR_LEVEL])
+                    $message,
+                    implode(PHP_EOL, $errors[$this->errorLevel])
                 );
             }
         } catch (ServiceMismatchException $e) {
